move interactiontype enum
