+++ api/src/main/java/de/oliver/fancynpcs/api/utils/InteractionType.java
@@ -1,0 +1,10 @@
+package de.oliver.fancynpcs.api.utils;
+
+public enum InteractionType {
+    LEFT_CLICK,
+    RIGHT_CLICK,
+    /**
+     * {@link InteractionType#CUSTOM InteractionType#CUSTOM} represents interactions invoked by the API.
+     */
+    CUSTOM
+}
